release optimizations

+++ day04/day04.hpp
@@ -36,7 +36,7 @@ private:
 
         {  // Part 2
             int sum = 0;
-            for (int i = 0; i < cards.size(); i++) {
+            for (size_t i = 0; i < cards.size(); i++) {
                 for (int j = 0; j < cards[i].matches; j++) {
                     cards[i + 1 + j].count += cards[i].count;
                 }
